Compute Posterior Summaries From Poisson Mash Fit

Compute posterior summaries of the given contrasts of the matrix of effects based on the poisson mash fit. Mixture components with very small posterior weights are ignored in the posterior calculations.

Note this is an internal function which users might not want to call directly.

pois_mash_posterior(
  data,
  s,
  mu,
  psi2,
  bias = matrix(0, nrow(data), ncol(data)),
  wlist,
  Ulist,
  ulist,
  ulist.epsilon2 = rep(1e-08, length(ulist)),
  zeta,
  thresh = 1/(500 * ncol(zeta)),
  C = diag(ncol(data)) - 1/ncol(data),
  res.colnames = paste0(colnames(data), "-mean"),
  posterior_samples = 0,
  median_deviations = FALSE,
  seed = 1
)

Arguments

data

J x R matrix of counts collapsed over conditions, with features as rows and conditions as columns.

s

Numeric vector of length R adjusting for sequencing depth of each of R conditions.

mu

J x R matrix of gene-specific means (R conditions are assumed to belong to M subgroups, so each row should have at most M distinct values).

psi2

Vector of length J giving the gene-specific dispersion parameters.

bias

J x R matrix of bias caused by unwanted variation. Default is a matrix of all zeros.

wlist

Numeric vector of length L specifying the scaling factors for the prior covariance matrices.

Ulist

List of H full-rank covariance matrices

ulist

List of G numeric vectors each of which forms a rank-1 covariance matrix.

ulist.epsilon2

Numeric vector of length G added to the diagonals of each rank-1 prior covariance matrix.

zeta

J x K matrix of posterior weights, where K = L*(H + G) is the number of prior mixture components.

thresh

Posterior weights thresholds. Below this threshold, the mixture components are ignored in the posterior calculations.

C

Q x R matrix of contrasts for effects.

res.colnames

Character vector of length Q giving the names of the contrasts.

posterior_samples

The number of samples to be drawn from the posterior distribution of each effect.

median_deviations

Logical scalar indicating whether to calculate posterior summary of deviation of condition-specific effects relative to the median over all conditions.

seed

a random number seed to use when sampling from the posteriors. It is used when posterior_samples > 0 or median_deviations = TRUE.

Value

The return value is a list with the following components:

PosteriorMean

J x Q matrix of posterior means.

PosteriorSD

J x Q matrix of posterior standard deviations.

ZeroProb

J x Q matrix in which each entry is the posterior probability of the true effect being zero.

NegativeProb

J x Q matrix in which each entry is the posterior probability of the true effect being negative.

lfsr

J x Q matrix of local false sign rate estimates.

PosteriorSamples

J x R x posterior_samples array of samples of effects, if posterior_samples > 0.

beta_median_dev_post

a list containing posterior summary of deviation of effects relative to the median, if median_deviations = TRUE.